Extract common logic for email validation. Migrate from chars to code points
